2012-09-03  Basile Starynkevitch  <basile@starynkevitch.net> 	* melt-runtime.h (melt_string_length): Add cast.
	{{Regenerate}}

	* melt/generated/meltrunsup-inc.c: Regenerate.
	* melt/generated/meltrunsup.h: Regenerate.
	* melt/generated/warmelt-base+01.c: Regenerate.
	* melt/generated/warmelt-base+02.c: Regenerate.
	* melt/generated/warmelt-base+03.c: Regenerate.
	* melt/generated/warmelt-base+meltdesc.c: Regenerate.
	* melt/generated/warmelt-base+melttime.h: Regenerate.
	* melt/generated/warmelt-base.c: Regenerate.
	* melt/generated/warmelt-debug+meltdesc.c: Regenerate.
	* melt/generated/warmelt-debug+melttime.h: Regenerate.
	* melt/generated/warmelt-debug.c: Regenerate.
	* melt/generated/warmelt-first+meltdesc.c: Regenerate.
	* melt/generated/warmelt-first+melttime.h: Regenerate.
	* melt/generated/warmelt-first.c: Regenerate.
	* melt/generated/warmelt-genobj+04.c: Regenerate.
	* melt/generated/warmelt-genobj+05.c: Regenerate.
	* melt/generated/warmelt-genobj+meltdesc.c: Regenerate.
	* melt/generated/warmelt-genobj+melttime.h: Regenerate.
	* melt/generated/warmelt-genobj.c: Regenerate.
	* melt/generated/warmelt-macro+meltdesc.c: Regenerate.
	* melt/generated/warmelt-macro+melttime.h: Regenerate.
	* melt/generated/warmelt-macro.c: Regenerate.
	* melt/generated/warmelt-modes+meltdesc.c: Regenerate.
	* melt/generated/warmelt-modes+melttime.h: Regenerate.
	* melt/generated/warmelt-modes.c: Regenerate.
	* melt/generated/warmelt-normal+meltdesc.c: Regenerate.
	* melt/generated/warmelt-normal+melttime.h: Regenerate.
	* melt/generated/warmelt-normal.c: Regenerate.
	* melt/generated/warmelt-normatch+04.c: Regenerate.
	* melt/generated/warmelt-normatch+meltdesc.c: Regenerate.
	* melt/generated/warmelt-normatch+melttime.h: Regenerate.
	* melt/generated/warmelt-normatch.c: Regenerate.
	* melt/generated/warmelt-outobj+meltdesc.c: Regenerate.
	* melt/generated/warmelt-outobj+melttime.h: Regenerate.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/melt-branch@190894 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/melt-runtime.h
@@ -1548,7 +1548,8 @@ melt_string_nth (melt_ptr_t v, int rk)
       unsigned slen = ((struct meltstring_st *) v)->slen;
       if (rk < 0) 
 	rk += (int) slen;
-      if (rk >= 0 && rk < slen) return ((struct meltstring_st *) v)->val[rk];
+      if (rk >= 0 && rk < (int)slen) 
+	return ((struct meltstring_st *) v)->val[rk];
     };
   return 0;
 }
